wip: refactor move parser-schema
