+++ .changeset/red-eagles-whisper.md
@@ -1,0 +1,7 @@
+---
+'@journeyapps/db': minor
+'@journeyapps/parser-schema': minor
+
+---
+
+Refactor Primitives to use abstraction instead of prototypes
