+++ journeyapps-evaluator/src/TypeInterface.ts
@@ -1,12 +1,12 @@
 export interface TypeInterface {
   name: string;
   isPrimitiveType: boolean;
+  options: object;
+  objectType?: TypeInterface;
+  getType(expression: string): TypeInterface | null;
   getVariable(expression: string);
   getVariableTypeAndNameWithParent(expression: string);
-  getType(expression: string): TypeInterface | null;
-  toJSON(): any;
   format(value: any, format?: string): string;
-  options: object;
   stringify(): string;
-  objectType?: TypeInterface;
+  toJSON(): any;
 }
