+++ journeyapps-parser-schema/src/index.ts
@@ -1,12 +1,12 @@
-export * from './ArrayType';
-export * from './EnumOption';
-export * from './NestedType';
-export * from './ObjectType';
+export * from './types/ArrayType';
+export * from './schema/EnumOption';
+export * from './types/NestedType';
+export * from './types/ObjectType';
 export * from './primitives';
-export * from './QueryType';
-export * from './Relationship';
-export * from './Schema';
-export * from './schemaParser';
-export * from './schemaSerializer';
-export * from './Type';
-export * from './Variable';
+export * from './types/QueryType';
+export * from './schema/Relationship';
+export * from './schema/Schema';
+export * from './schema/schemaParser';
+export * from './schema/schemaSerializer';
+export * from './types/Type';
+export * from './schema/Variable';
