+++ journeyapps-parser-schema/src/schema/EnumOption.ts
