+++ journeyapps-parser-schema/src/schema/Schema.ts
@@ -4,11 +4,11 @@
 import { Version, DEFAULT as DEFAULT_VERSION } from '@journeyapps/parser-common';
 import * as parser from './schemaParser';
 import { Variable } from './Variable';
-import { QueryType } from './QueryType';
-import { ArrayType } from './ArrayType';
-import { ObjectType } from './ObjectType';
-import { Type } from './Type';
-import { primitives, primitive } from './primitives';
+import { QueryType } from '../types/QueryType';
+import { ArrayType } from '../types/ArrayType';
+import { ObjectType } from '../types/ObjectType';
+import { Type } from '../types/Type';
+import { primitives, primitive } from '../primitives';
 import { Relationship } from './Relationship';
 import { ValidationError } from '@journeyapps/core-xml';
 import { XMLError, XMLElement } from '@journeyapps/domparser/types';
