+++ journeyapps-parser-schema/src/types/ArrayType.ts
