+++ journeyapps-parser-schema/src/types/ObjectType.ts
@@ -1,11 +1,11 @@
 import { Type } from './Type';
 import { FormatString } from '@journeyapps/evaluator';
-import { Variable } from './Variable';
-import { Relationship } from './Relationship';
-import { Schema } from './Schema';
+import { Variable } from '../schema/Variable';
+import { Relationship } from '../schema/Relationship';
+import { Schema } from '../schema/Schema';
 import { XMLElement } from '@journeyapps/domparser/types';
-import { primitive } from './primitives';
-import { ModelIndex } from './ModelIndex';
+import { primitive } from '../primitives';
+import { ModelIndex } from '../schema/ModelIndex';
 
 export class ObjectType extends Type {
   displayFormat: FormatString | null;
