+++ journeyapps-parser-schema/src/types/QueryType.ts
